+++ src/decorator_trait.jl
@@ -282,6 +282,8 @@ end
 # Introduce Deco Trait | automatic foward | fallback
 @trait_function get_vectors(M::AbstractDecoratorManifold, p, B::AbstractBasis)
 
+@trait_function has_components(M::AbstractDecoratorManifold)
+
 @trait_function injectivity_radius(M::AbstractDecoratorManifold)
 function injectivity_radius(
     ::TraitList{IsIsometricEmbeddedManifold},