+++ NEWS.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.15.19] 20/10/2024
+
+### Changed
+
+* make `has_components` introduced in the last version a decorator trait function.
+
 ## [0.15.18] 18/10/2024
 
 ### Added